clay_print_tap.c: print the TAP test plan at the end instead of beginning
Since we may not know how many tests will be run, for example if the user
selected a subset of test suites to run on the command-line, we cannot
print the TAP plan at the beginning.  Instead, wait till the end and print
out the plan based on the count of tests that were run.

This will trigger a bug in the perl TAP::Parser module version 3.17 if any
of the tests fail and any YAML output is produced, but it can't be helped.
The bug should be fixed in version 3.22.