Add flag to exit on first error
This adds a "-Q" (quit) flag to cause the test suite to exit
after the first error occurs.  This shouldn't be the default
behavior, but sometimes when I really mess up the tests, I
just want to see the first problem and fix it, knowing that
that will probably fix the next 1000 errors, too.