Allow -s to specific individual test names
Previously -s had to be followed by the name of a test suite or a
prefix of the name of a set of test suites to be run.  Now it can
take an argument all the way down to the individual test level and
force just that single test to be executed.