Convert assert_equal to stdarg for formats
It is nice to be able to override the numeric formatting when
using the cl_assert_equal helpers (in particular, I wanted to be
able to have octal formatting for numbers).  To allow this, I
consolidated the different clar__assert_equal_X helpers into a
single one that uses varargs and examines a new "format" input
that tells what format is expected for the two values to be
compared.

To test this, I expanded the sample tests to include broader
coverage and I made sure that each example test included both one
or more passing tests and a failing test.

Additionally, when the snprintf_eq macro was fixed to use variadic
arguments, the name no longer made sense (since it was no longer
being used just to print equality of two values).  I renamed the
whole macro to the more generic p_snprintf to indicate that it is
just a simple "portable" wrapper to snprintf.