Introduce assertions for breakage
Next to explicitly setting a breakage inside of a test with `cl_break`,
the most common scenario will most likely be to have known breakage on a
condition. This patch introduces a new set of macros `cl_must_break` and
`cl_check_break`. These macros cause a test to either be set to "broken"
if the assertion holds or, if it does not, to "failure". Like this, the
test can notify on both conditions:

- the system under test is still broken

- the system under test has been fixed and is not broken anymore,
  resulting in a failure

Note that a failure is the only reasonable thing to do. The authors of
the test suite certainly want to be notified as soon as known breakages
get fixed such that they can adjust tests and potentially e.g. remove
TODO items or bug tickets associated with the previous breakage.