Add status to error structure
Printing out errors can be based on the type of error status. While
there is currently no status besides test failures which is to be
printed, we will introduce a failure type "Unbroken" soon. As such, add
the error status to the structure so that we're able to distinguish
errors and print them out differently.