clar: introduce function to abort execution
In almost all cases, a call to `clar_print_onabort()` is immediately
followed by a call to `exit(-1)`. Add a new function `clar_abort()` that
combines these two steps into one and convert callsites to use it.

While at it, adapt the error messages to be consistent with trailing
punctuation and newlines as well as more informative by including both
potential arguments and the errno.