clar/sandbox: convert to use `clar_abort()`
The function `clar_sandbox()` can fail either because it failed to build
a sandbox path or because it failed to change directories into it. The
error reporting happens at the callsite of this function though, where
we are unable to distinguish those two root causes.

Adapt the code to instead perform error handling via `clar_abort()` in
`clar_sandbox()` itself. This gives us direct access to the root cause.
Callers do not need to check for errors anymore.