clar/summary: convert to use `clar_abort()`
When initializing a summary fails we first call `perror()` followed by
`clar_abort()`. This is somewhat broken because the second function call
will always claim that opening the file has failed, even though the
failure may actually be that we are out of memory.

Fix this by using `clar_abort()` in `clar_summary_init()` directly,
which gives direct access to the failures' root cause. Callers do not
need to check for errors anymore.