clar: stop using `assert()`
We have a single callsite of `assert()` remaining in our codebase that
checks for unknown command line arguments. This is somewhat broken
though as asserts may not be compiled depending on how the code is
compiled. Consequently, we may not abort properly on invalid arguments
in some scenarios.

Fix this issue by using `clar_abort()` instead.