clar: allow overriding current file, line and function
When raising an error we print the current file, line and function where
the error was raised. This is of course quite helpful to users, but it
stands in the way of testing the clar because they make the output of
executables dependent on the developer's setup.

Add the ability to override these with fixed constants via a new
`CLAR_SELFTEST` macro.