test: exercise the clar itself
While we have had the "clar_test" binary for a very long time already,
it was more for demonstration purposes than for real testing of the
clar. It could serve as sort of a smoke test if the developer remembered
to execute it, but that's about it.

In the preceding commits we have moved the "clar_test" into a separate
directory, renamed it to "selftest_suite" and made its output
deterministic. We can thus now wire up proper testing of the clar by
exercising the "selftest_suite" binary via a new "selftest" binary.
These tests run the "selftest_test" binary with various different
arguments and check both its return code as well as its output.