test: drop global test counter
The global test counter has been added as a demonstration for how it is
possible to modify global resources. Now that we have moved examples
into their own directory though it no longer serves that purpose.

In fact, it now stands in our way as we are about to introduce testing
of the clar itself with various different options. But as we assert that
the counter is "8" after all tests have ran, we wouldn't be able to
exclude some tests from running.

Drop the counter altogether to prepare for the change.