clar: simplify how we handle stat(3P) on Win32
The Windows platform has multiple different definitions for stat(3P) and
its `struct stat` with different integer widths for both the filesize
and the timestamps. For our usecase, we really only need to care whether
we end up using `_stat()` or `stat()`, which require us to use either
`struct _stat` or `struct stat`, respectively.

The way we handle this in "clar.c" is somewhat convoluted though because
we select the function and structure to use independently of each other.
This can cause us to pick the wrong combination of features in some
environments, like mingw-w64.

Refactor the code so that both function and struct get set up under the
same condition, thus fixing the issue.