clar: support compilation without support for `wchar_t`
Some implementations of libc do not have proper support for `wchar_t`.
One such library is uClibc, which only has optional support for this
type. When not enabled, then the <wchar.h> header will be a mere stub
that only contains a typedef for `wchar_t`, but nothing else.

Drop our use of `wchar_t` on such platforms.

Reported-by: Bagas Sanjaya <bagasdotme@gmail.com>
Helped-by: Edgar Bonet <bonet@grenoble.cnrs.fr>