invoke: preserve line numbers during helper methods
When invoking a helper function (or "auxiliary method", in the
documentation), we lose the file/function/line information of the
current call point. This means that we attribute failures to the helper
function instead of the actual test.

Provide a `cl_invoke` function that will preserve the current state.
This allows callers to use the simpler "clar style" as documented, but
get the correct failure location.