add test_start print callback
Refactor the `ontest` callback (which is implicitly test _finished_)
into a test started and test finished callback. This allows printers to
show the test name (at start) and its conclusion in two steps, which is
advantageous for users to see the current test during long-running test
executions.

In addition, rename `onsuite` to `suite_start` for consistency.