multiple runs
Allow tests to specify that they should have multiple runs. These runs
all occur within a single initialization and cleanup phase, and is
useful for repeatedly testing the same thing as quickly as possible. The
time for each run is recorded, which may be useful for benchmarking that
test run.