in benchmark mode run iteratively
In benchmark mode, when the number of runs was not explicitly specified
in the test itself, run a reasonably number of iterations. We do this by
measuring one run of the test, then using that data to determine how
many iterations we should run to fit within 3 seconds. (With a minimum
number of iterations to ensure that we get some data, and a maximum
number to deal with poor precision for fast test runs.)

The 3 second number, and 10 iteration minimum, were chosen by consulting
the hyperfine defaults.