introduce a reset function
For multi-run tests (benchmarks), we introduce a `reset` function. By
default, between each run of a test, the initialization will be called
at startup, and the cleanup will be called at finish. A benchmark may
wish to set up multi-run state at the beginning of the invocation (in
initialization), and keep a steady state through all test runs. Users
can now add a `reset` function so that initialization occurs only at the
beginning of all runs.