Using test names as sandbox dirs is now optional
Given that test names can be long _and_ some platforms insist upon a
short maximum path length (eg, Windows), we may not want to
indiscriminately use path names as the sandbox directory name.

However, this has high utility in some situations, so allow it as an
opt-in. By default, we'll just use an 8 hexadigit "random" string.
(We'll use the rand(3) mechanism, which is not be particularly random,
but since the sandbox directories are sequestered in a unique tempdir,
a clar process need only attempt to avoid collisions between itself,
not with other processes.)