indent all lines in error messages and descriptions
When printing error messages and descriptions indent all the lines
rather than just the first line. When the tests are run with --tap this
ensures that the yaml snippet in the TAP output is valid.

The new test uses cl_must_pass() but multi-line error descriptions are
also likely to be created by cl_assert_equal_s().