avoid passing NULL to snprintf()
Passing a NULL pointer as the argument to a "%s" format specifier is
undefined according to the C standard. In practice glibc and some other
implementations accept NULL and print "(null)" but such behavior cannot
be relied upon by portable code. To avoid any ambiguity with the string
"NULL" NULL pointers are printed "NULL" without the single quotes that
surround the text of non-NULL strings.