clar: adapt "-q" to decrement verbosity
The clar understands both "-v" and "-q". Both of these arguments
effectively increase or decrease verbosity, but right now their
respective effects are somewhat different from one another. While "-v"
makes us print more information, "-q" only skips over successful tests
without modifying the verbosity itself.

This behaviour is somewhat inconsistent and confusing, and the resulting
implementation is more complex than it really ought to be. This is
demonstrated by the fact that we have several bugs when it comes to the
verbosity, all of which seem to be consequences of this complexity.

Refactor the code so that we only have a single "verbosity". "-v"
increases verbosity, whereas "-q" decreases verbosity. When verbosity is
negative, we start to reduce our verbosity, whereas positive values will
start to increase our verbosity.

This refactoring is a first step towards improving quiet output over the
next couple of commits.