clar/print: handle negative verbosity in format implementations
Both `clar_print_ontest()` and `clar_print_onsuite()` are only called
when we aren't asked to be quiet. This results in buggy behaviour
though, where we for example don't print tests at all with the TAP
format, which is invalid.

Refactor the code so that the format implementations themselves can
decide whether or not to skip writing output based on verbosity. For now
the behaviour remains the same, but we will iterate on the respective
formats over the next couple of commits.