clar/print: fix TAP output in quiet mode
The TAP format we print when asked to be quiet is broken because we
don't the "ok/not ok" line for any of the individual tests. This does
not follow the specification and would thus confuse parsers.

Fix the issue by printing tests even when asked to be quiet. To have the
"-q" flag still have an effect we instead skip the reason why the test
has failed.