clar/print: improve clap format in quiet mode
When asked to be quiet we don't print the status for each of the
individual tests (the "...FF..F" indicator). But we still end up
printing the legend explaining what each of the indicator means as well
as how many test suites we have loaded, both of which feel unnecessarily
verbose.

Stop printing this information in quiet mode. The end result is that we
only end up printing failures, but nothing else.