clar/print: drop unused suite names
When printing the initial status we pass the number of tests, the number
of suites and the suite names to `clar_print_init()`. But while the
former two pieces of information do get used when using the Clap format,
we only ever pass an empty string as suite name.

Besides being unnecessary, it also has the consequence that we print an
empty space, as the final "%s" in "Loaded %d suites: %s\n" is always going
to be the empty string.

Remove the unused parameter and drop the trailing space. Adapt our tests
accordingly.