generate.py: allow generating suite with single file
Usually, one is expected to pass the path to a directory to
"generate.py". The script will then load all ".c" files in that
directory and extract unit tests from them. There are use cases though
where one may want to only scan a single source file. While it is of
course possible to put that file into its own directory, that may be
unwieldy.

Allow passing file names to "generate.py". If so, the script will treat
this file as the only module for the generated test suite.

There is one caveat here: we know derive the location of the generated
"clar.suite" file from the directory. This doesn't really work well
anymore when passed a single file. While it is in theory possible to
just use that file's directory, in practice this may be a bit on the
surprising side. Instead, we mandate that the `--output` option must be
passed by the user when given a file.