test: allow for multiple test suites
Right now, our self-test suite only has a single child test suite that
it knows to execute. This has the downside that any new test we add to
that child test suite requires us to make a bunch of modifications to
all of our test assumptions.

Furthermore, we don't have an easy way to add tests that require us to
do some post-processing of the output generated by the test suite. While
that could (and will) be addressed by splitting up the responsibility of
`run()`, the bigger problem is still that such a test would be executed
by default when running all tests.

Refactor our selftests so that we can have multiple different child test
suites.