clar: use "%p" to print mismatching pointers
In 8807789 (clar: remove unportable use of `uintptr_t`, 2025-09-18) we
have adapted how we print pointers in a failing assert so that we use
`uintmax_t` instead of `uintptr_t`. That doesn't work on 32 bit systems
though:

    ::error file=clar.c,line=879::clar.c:879:8: cast from pointer to integer of different size [-Werror=pointer-to-int-cast]
      879 |        (uintmax_t)p1, (uintmax_t)p2);
          |        ^

So we cannot use either `uintmax_t` nor `uintptr_t`.

Let's bite the bullet and use "%p" as a formatting directive instead so
that we can print the pointer values without a cast.

The downside of this is that the output is now platform-dependent, which
is okay -- after all, pointers are platform-dependent by definition. We
thus extract the pointer-related tests into a separate test suite and
only perform check for substrings to exist.