tap: fix yaml delimiter
The TAP 13 spec [1] states that the trailing delimiter for yaml
diagnostics is "\s+...\n", not "\s+---\n". This causes prove to fail
to parse the output of git's unit tests when they fail. NB If we
ever want to change to using TAP 14 then we'll need to change the
indentation as it requires a strict two space indentation.

[1] https://testanything.org/tap-version-13-specification.html

Reported-by: Jeff King <peff@peff.net>