clar: introduce `clar_failf()`
While we already have `clar_fail()`, this function does not allow the
caller to provide a printf(3p)-style formatting string. This makes it
somewhat hard at times to give a proper explanation _why_ a specific
test has failed.

Introduce `clar_failf()` to plug this gap. Note that the core of this
new functionality is implemented in `clar__failv()`, which receives a
`va_list` as input. This indirection isn't needed right now, but we will
add more callers in subsequent commits.