clar: avoid repetition in integer comparison functions
The error message that we use for all the integer comparison functions
is repeated for every macro. This was done because we reused the same
infra for both deprecated comparison functions (`cl_assert_equal_i()`)
and the new replacements (`cl_assert_eq_i()`), and the resulting error
messages are expected to be different there.

That has changed in the preceding commit though, where we started to
open-code the deprecated functions. Consequently, we can now deduplicate
the error message for the new comparison functions.

Do so.