clar: define CLAR_MAX_PATH even w/o PATH_MAX
PATH_MAX is an optional constant in POSIX, and at least on GNU/Hurd
it is not provided on purpose.

Define CLAR_MAX_PATH to a static value in case PATH_MAX is not available,
using the same value used for the "Windows with long paths" case. In the
end, 4096 should not be a problem for a test framework.